% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ampkTorDomain}
\alias{ampkTorDomain}
\title{Protein domain annotations of AMPK-TOR data}
\format{A data frame with 5 columns and 133489 rows:
\itemize{
    \item{seedID}{Gene ID & its ortholog, e.g. 
    "ampk_ACACA#ampk_ACACA|ANOGA@7165@1|Q7PQ11|1"}
    \item{orthoID}{Seed/Ortholog ID, e.g. 
    "ampk_ACACA|ANOGA@7165@1|Q7PQ11|1"}
    \item{feature}{Annotated feature/domain, e.g. "pfam_CPSase_L_D2"}
    \item{start}{Start position, e.g. "260"}
    \item{end}{End position, e.g. "462"}
}}
\usage{
data(ampkTorDomain)
}
\description{
Write the description here (what are the scores, how many taxa, how many 
proteins)
}
\keyword{datasets}
