% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ampkTorFasta}
\alias{ampkTorFasta}
\title{FASTA sequences of AMPK-TOR data}
\format{A AAStringSet object (package Biostrings)}
\usage{
data(ampkTorFasta)
}
\description{
A list of fasta sequences (AAStringSet objects) for 6535 proteins of the 
phylogenetic profiles. They include 147 human proteins that are part of the 
human AMPK-TOR pathway and 6388 orthologous proteins in other 82 species.
}
\keyword{datasets}
