% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arthropodaDomain}
\alias{arthropodaDomain}
\title{Protein domain annotations of Arthropoda data}
\format{A data frame with 8 columns and 866912 rows:
\itemize{
    \item{seedID}{Gene ID & its ortholog, e.g. 
    "136365at6656#136365at6656|ANOGA@7165@1|Q7QC64|1"}
    \item{orthoID}{Seed/Ortholog ID, e.g. 
    "136365at6656|ANOGA@7165@1|Q7QC64|1"}
    \item{length}{Seed/Ortholog length, e.g. "142"}
    \item{feature}{Annotated feature/domain, e.g. "pfam_Ribosomal_L27"}
    \item{start}{Start position, e.g. "26"}
    \item{end}{End position, e.g. "106"}
    \item{weight}{Feature weight, e.g. "NA"}
    \item{path}{Feature is used for comparison (Y/N), e.g. "Y"}
}}
\usage{
data(arthropodaDomain)
}
\description{
Data frame containing the domain annotation for all proteins in phylogenetic 
profiles of the arthropoda dataset.
}
\keyword{datasets}
