% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arthropodaFasta}
\alias{arthropodaFasta}
\title{FASTA sequences of Arthropoda data}
\format{A AAStringSet object (package Biostrings)}
\usage{
data(arthropodaFasta)
}
\description{
Write the description here (how many species, how many sequences, where can 
one get the data)
}
\keyword{datasets}
