% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arthropodaPhyloProfile}
\alias{arthropodaPhyloProfile}
\title{Phylogenetic profiles of Arthropoda taxon set}
\format{A data frame with 5 columns and 54656 rows:
\itemize{
    \item{geneID}{Gene ID, e.g. "97421at6656"}
    \item{ncbiID}{Taxonomy ID of the search species, e.g. "ncbi9598"}
    \item{orthoID}{Ortholog ID, e.g. "97421at6656|PANTR@9598@1|H2QTF9|1"}
    \item{FAS_F}{Forward FAS score, e.g. "0.6872810"}
    \item{FAS_B}{Backward FAS score, e.g. "0.9654661"}
}}
\usage{
data(arthropodaPhyloProfile)
}
\description{
Write the description here (what are the scores, how many taxa, how many 
proteins - what are those proteins)
}
\keyword{datasets}
