% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ampkTorFasta}
\alias{ampkTorFasta}
\title{FASTA sequences of AMPK-TOR data}
\format{A AAStringSet object (package Biostrings)}
\usage{
data(ampkTorFasta)
}
\description{
Write the description here (how many species, how many sequences, where can 
one get the data)
}
\keyword{datasets}
