% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ampkTorPhyloProfile}
\alias{ampkTorPhyloProfile}
\title{Phylogenetic profiles of AMPK-TOR taxon set}
\format{A data frame with 5 columns and 6535 rows:
\itemize{
    \item{geneID}{Gene ID, e.g. "ampk_ACACA"}
    \item{ncbiID}{Taxonomy ID of the search species, e.g. "ncbi284812"}
    \item{orthoID}{Ortholog ID, e.g. "ampk_ACACA|SCHPO@284812@1|P78820|1"}
    \item{FAS_F}{Forward FAS score, e.g. "0.9884601"}
    \item{FAS_B}{Backward FAS score, e.g. "0.9907436"}
}}
\usage{
data(ampkTorPhyloProfile)
}
\description{
Write the description here (what are the scores, how many taxa, how many 
proteins - what are those proteins)
}
\keyword{datasets}
