% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arthropodaFasta}
\alias{arthropodaFasta}
\title{FASTA sequences of Arthropoda data}
\format{A AAStringSet object (package Biostrings)}
\usage{
data(arthropodaFasta)
}
\description{
A list of fasta sequences (AAStringSet objects) for 1011 BUSCO ortholog 
groups of the phylogenetic profile. It has in total 54488 protein sequences.
}
\keyword{datasets}
