% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncbiTaxonNamesFull}
\alias{ncbiTaxonNamesFull}
\title{NCBI Taxonomy full data set}
\format{A data frame with 4 columns and 2110551 rows:
\itemize{
    \item{ncbiID}{ e.g. "10090"}
    \item{fullName}{ e.g. "Mus musculus"}
    \item{rank}{ e.g. "species"}
    \item{parentID}{ e.g. "862507"}
}}
\usage{
data(ncbiTaxonNamesFull)
}
\description{
A list of all NCBI taxon IDs, taxon names, their systematic taxonomy rank and
IDs of their next ranks (parent IDs)
}
\keyword{datasets}
