% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloProfileData.R
\docType{package}
\name{PhyloProfileData}
\alias{PhyloProfileData}
\alias{PhyloProfileData-package}
\title{Data package for phylogenetic profile analysis using PhyloProfile tool}
\description{
The PhyloProfileData package provides a collection of datasets 
to accompany the R package PhyloProfile, where they are used to illustrate 
how to run PhyloProfile and analyse its results. Briefly, it contains the 
phylogenetic profiles, the fasta sequences and the domain annotations for two
experimental data sets, including human AMPK-TOR pathway proteins and BUSCO 
Arthropoda proteins.
}
\details{
More details of the datas can be found in the vignette 
"PhyloProfileData".
}
